[FLINK-14588][hive] Support Hive version 1.0.0 and 1.0.1
