+++ flink-connectors/flink-connector-hive/pom.xml
@@ -666,7 +666,23 @@ under the License.
 	</build>
 
 	<profiles>
-		<!-- Activate this profile with -Phive-1.2.1 to build and test against hive-1.2.1 -->
+		<!-- Activate these profiles with -Phive-x.x.x to build and test against different Hive versions -->
+		<profile>
+			<id>hive-1.0.0</id>
+			<properties>
+				<hive.version>1.0.0</hive.version>
+				<hivemetastore.hadoop.version>2.6.5</hivemetastore.hadoop.version>
+				<hiverunner.version>3.1.1</hiverunner.version>
+			</properties>
+		</profile>
+		<profile>
+			<id>hive-1.0.1</id>
+			<properties>
+				<hive.version>1.0.1</hive.version>
+				<hivemetastore.hadoop.version>2.6.5</hivemetastore.hadoop.version>
+				<hiverunner.version>3.1.1</hiverunner.version>
+			</properties>
+		</profile>
 		<profile>
 			<id>hive-1.1.1</id>
 			<properties>
