+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/connectors/hive/HiveRunnerShimLoader.java
@@ -37,6 +37,8 @@ public class HiveRunnerShimLoader {
 		String hiveVersion = HiveShimLoader.getHiveVersion();
 		return hiveRunnerShims.computeIfAbsent(hiveVersion, v -> {
 			switch (v) {
+				case HiveShimLoader.HIVE_VERSION_V1_0_0:
+				case HiveShimLoader.HIVE_VERSION_V1_0_1:
 				case HiveShimLoader.HIVE_VERSION_V1_1_0:
 				case HiveShimLoader.HIVE_VERSION_V1_1_1:
 				case HiveShimLoader.HIVE_VERSION_V1_2_0:
